backend/requirements.txt
blinker==1.8.2
click==8.1.7
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.0.3
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
openpyxl==3.1.5
tqdm==4.66.6
Werkzeug==3.1.1
